:hocho: extract episodes list to reduce complexity
